Ensure widget styles apply everywhere
Some themes do not use the class 'widget' for their widgets by default. (ex: BeaverBuilder theme which uses fl-widget), so either the stylesheets need to be updated to not rely upon the 'widget' class, or this change needs to be implemented to ensure the widget class is used on all wp-listings widgets.